+++ test/src/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import './commandlinker/linker.spec';
+import './commandlinker/commandlinker.spec';
 
 import './common/activitymonitor.spec';
 import './common/instancetracker.spec';
